requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
Flask==3.1.0
huggingface_hub==0.24.5
numpy==1.26.4
onnxruntime_gpu==1.18.1
pandas==2.2.3
Pillow==11.1.0
torch==2.4.0+cu124